requirements.txt
djangorestframework==3.11.2
django
pyqrcode
pypng